Adjusted style form added a side menu navigation. Added a triangle shape to open the side menu navigation.
Adjusted the login form and register form so that it is enclosed within the side menu navigation.
the elements are currently not entirely cnfigured to display the login form's style.
Also the elements for the register page are not configured to dispaly the register page.

Still some other code needs to be written.