adjusted controllers routes for handling user login and registration. At code level added a limit of only allowing 1 user to have 1 account per email address.
Modified the authenticate and compare functions to allow for comparing the user input with the database stored information.

Modified Profile Completion Routes to handle user profile completion.

Adjusted the JWT token generation to generate token based authentication and map to each of the unique users.

Changed the functions handling the authentication flow in a separate file and exported the modules.