Update index.html
removed mailto and ftp protocols as the products would only be updated via the https transfer and related list of products would be updated as in stock or out of stock via https and not http. So, http should be removed later on only via secure protocols as encrypted data.



During live development it would be required to have the actual VPN as the connection not to be hijacked by some idiots that would get the data instead